refactor: BaseEntity 통일 #12
