+++ src/main/java/com/main/writeRoom/aws/s3/Uuid.java
@@ -1,6 +1,6 @@
 package com.main.writeRoom.aws.s3;
 
-import com.main.writeRoom.common.BaseEntity;
+import com.main.writeRoom.domain.common.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
