+++ src/main/java/com/main/writeRoom/domain/Room.java
@@ -1,8 +1,7 @@
 package com.main.writeRoom.domain;
 
-import com.main.writeRoom.common.BaseEntity;
-
 import com.main.writeRoom.domain.Challenge.ChallengeRoutine;
+import com.main.writeRoom.domain.common.BaseEntity;
 import com.main.writeRoom.domain.mapping.RoomParticipation;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
