+++ src/main/java/com/main/writeRoom/domain/mapping/RoomParticipation.java
@@ -2,9 +2,9 @@ package com.main.writeRoom.domain.mapping;
 
 import static com.main.writeRoom.apiPayload.status.ErrorStatus.AUTHORITY_TYPE_ERROR;
 
-import com.main.writeRoom.common.BaseEntity;
 import com.main.writeRoom.domain.Room;
 import com.main.writeRoom.domain.User.User;
+import com.main.writeRoom.domain.common.BaseEntity;
 import com.main.writeRoom.handler.RoomHandler;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
